requirements.txt
requests
GitPython